requirements.txt
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy<2.0
rapidfuzz
opencv-python
opencv-contrib-python
cython
Pillow
pyyaml
requests
albumentations==1.4.10
albucore==0.0.13
python-docx==1.1.2
beautifulsoup4==4.12.3
pdf2image==1.17.0
paddlenlp==2.6.1
ppocrlabel
pandas
openpyxl